refactor: Convert config to feature
