+++ lib/mdigits.dart
@@ -2,4 +2,4 @@ library mdigits;
 
 export 'src/mdigits/mdigits_activity.dart';
 export 'src/models/mdigits_data.dart';
-export 'src/models/config.dart';
+export 'src/config/config.dart';
