+++ lib/src/mdigits/mdigits_activity.dart
@@ -3,7 +3,7 @@ import 'package:get/get.dart';
 import 'package:mdigits/src/data.dart';
 import 'package:mdigits/src/mdigits/mdigits_controller.dart';
 import 'package:mdigits/src/mdigits/mdigits_view.dart';
-import 'package:mdigits/src/models/config.dart';
+import 'package:mdigits/src/config/config.dart';
 import 'package:mdigits/src/models/mdigits_data.dart';
 
 /// Present the appropriate screen based on the curren step
