+++ lib/src/mdigits/mdigits_view.dart
@@ -4,7 +4,7 @@ import 'package:mdigits/src/data.dart';
 import 'package:mdigits/src/end/end_view.dart';
 import 'package:mdigits/src/instructions/instructions_view.dart';
 import 'package:mdigits/src/mdigits/mdigits_controller.dart';
-import 'package:mdigits/src/models/config.dart';
+import 'package:mdigits/src/config/config.dart';
 import 'package:mdigits/src/response/response_view.dart';
 import 'package:mdigits/src/rest/rest_view.dart';
 import 'package:mdigits/src/stim/stim_view.dart';
